pyproject.toml
[tool.poetry]
name = "zodchy-notations"
version = "0.1.2"
description = ""
authors = ["Smairon <man@smairon.ru>"]
readme = "README.md"
repository = "https://github.com/smairon/zodchy-notations"

[tool.poetry.dependencies]
python = "^3.12"
zodchy = "^0.2.3"
python-dateutil = "^2.9.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"